+++ src/app/app.component.ts
@@ -1,14 +1,19 @@
 import { CartService } from './cart/services/cart.service';
-import { Component } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
+export class AppComponent implements AfterViewInit {
 
-export class AppComponent {
-  title = 'shop';
+  @ViewChild('appTitle') title!: ElementRef<HTMLHeadingElement>;
 
-  constructor(public cartService: CartService) {}
+  constructor(public cartService: CartService) {
+  }
+
+  ngAfterViewInit(): void {
+   this.title.nativeElement.textContent = 'Магазин "Найдется ВСЁ!"';
+  }
 }