+++ src/app/core/models/configModel.ts
@@ -1,0 +1,7 @@
+export class ConfigModel {
+  constructor(
+    public id: string,
+    public login: string,
+    public email: string) {
+  }
+}