+++ src/app/product/components/product-list/product-list.component.ts
@@ -1,3 +1,4 @@
+import { CartService } from './../../../cart/services/cart.service';
 import { Component, OnInit } from '@angular/core';
 import { ProductModel, ProductService } from '../../services/product.service';
 
@@ -9,10 +10,28 @@ import { ProductModel, ProductService } from '../../services/product.service';
 export class ProductListComponent implements OnInit {
 
   products: ProductModel[] = [];
-  constructor(private productService: ProductService) { }
+  sortByAscending: boolean = true;
+  sortByProperties: string[] = [];
+
+  constructor(private productService: ProductService,
+              private cartService: CartService) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
+  }
+
+  sortByAscendingChanged(value: boolean) {
+    this.sortByAscending = value;
+    console.log("!!!");
+  }
+
+  sortByPropertiesChanged(properties: string[]) {
+    this.sortByProperties = properties;
+  }
+
+  onAddToCart(product: ProductModel): void{
+    const transformedProduct = {...product, quantity: 1};
+    this.cartService.addProduct(transformedProduct);
   }
 
 }