+++ src/app/product/services/product.service.ts
@@ -9,9 +9,15 @@ export class ProductService {
 
   public getProducts(): ProductModel[] {
     const array: ProductModel[] = [];
-    array.push(new ProductModel());
-    array.push(new ProductModel());
-    array.push(new ProductModel());
+    array.push(new ProductModel(
+      'Product 1', 'Description for product 1', 10.0, true, Discount.No
+    ));
+    array.push(new ProductModel(
+      'Product 2', 'Description for product 2', 5.0, true, Discount.No
+    ));
+    array.push(new ProductModel(
+      'Product 3', 'Description for product 3', 20.0, true, Discount.No
+    ));
     return array;
   }
 }
@@ -19,22 +25,15 @@ export class ProductModel {
   constructor(
     public name: string = 'Name',
     public description: string = 'Description',
-    public price: number = 1.0,
+    public price: number = 0.0,
     public isAvailable: boolean = false,
-    public status: Status = Status.OnSale,
+    public status: Discount = Discount.No,
+    public quantity = 1
   ) {}
-
-  // public constructor() {
-  //   this.name = "Name";
-  //   this.description = "Description";
-  //   this.price = 1.0;
-  //   this.isAvailable = false;
-  //   this.status = Status.OnSale;
-  // }
-
 }
 
-export enum Status {
-  OnSale,
-  OnDiscount,
+export enum Discount {
+  No,
+  Five,
+  Ten
 }