+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [1.0.0] - 2017-06-20
+## [1.1.0] - 2021-05-23 
+### Added
+ - cart service
+ - appHighlight directive
+ - orderBy pipe 
+ - pipes added in product-list component and cart component
+
+## [1.0.0] - 2021-05-01
 ### Added
  - component Cart
  - component Product